requirements.txt
notion
python-dotenv
requests
django
djangorestframework
django-dotenv